+++ generators/app/templates/helpers/_grunt/imageSizeExport.js
@@ -1,0 +1,10 @@
+module.exports = {
+	folder: {
+		options: {
+			path: '<%= paths.src %>/assets/img/temp/pictures/**/*.jpg',
+			output: '<%= paths.src %>/assets/img/pictures.json',
+			categorizeBy: 'folders',
+			breakpointDelimiter: '--'
+		}
+	}
+};