+++ test/grunt-image-size-export-spec.js
@@ -1,17 +1,19 @@
 /*global describe, beforeEach, it*/
 'use strict';
 
+var fs = require('fs');
 var path = require('path');
 var helpers = require('yeoman-generator').test;
-var fs = require('fs');
+var assert = require('yeoman-generator').assert;
 var answers = require('../test_helpers/prompt-answer-factory')({
 	"gruntModules": [
-		"grunt-devtools"
+		"grunt-image-size-export"
 	]
 });
 
-
-describe('grunt-devtools', function () {
+describe('grunt-image-size-export', function () {
+	var srcPath = "resources/";
+	var helperPath = "helpers/";
 
 	beforeEach(function (done) {
 		helpers.run(path.join(__dirname, '../generators/app'))
@@ -25,6 +27,14 @@ describe('grunt-devtools', function () {
 	});
 
 	it('adds references to package.json', function () {
-		helpers.assertFile('package.json', /grunt-devtools/);
+		assert.fileContent('package.json', /grunt-image-size-export/);
+	});
+
+	it('creates helper files', function () {
+		assert.file(helperPath + "_grunt/imageSizeExport.js");
+	});
+
+	it('adds task to Gruntfile.js file', function () {
+		assert.fileContent("Gruntfile.js", /\'imageSizeExport\'/);
 	});
 });