+++ test/grunt-postcss-separator-spec.js
@@ -1,19 +1,20 @@
 /*global describe, beforeEach, it*/
 'use strict';
 
+var fs = require('fs');
 var path = require('path');
 var helpers = require('yeoman-generator').test;
 var assert = require('yeoman-generator').assert;
-var fs = require('fs');
 var answers = require('../test_helpers/prompt-answer-factory')({
-	"features": [
-		"installDocs"
+	"modules": [
+		"grunt-postcss-separator"
 	]
 });
 
-describe('grunt-styleguide', function () {
-	var helperPath = "helpers/";
+
+describe('grunt-postcss-separator', function () {
 	var srcPath = "resources/";
+	var helperPath = "helpers/";
 
 	beforeEach(function (done) {
 		helpers.run(path.join(__dirname, '../generators/app'))
@@ -27,23 +28,14 @@ describe('grunt-styleguide', function () {
 	});
 
 	it('adds references to package.json', function () {
-		assert.fileContent('package.json', /grunt-styleguide/);
+		assert.fileContent('package.json', /grunt-postcss-separator/);
 	});
 
 	it('creates helper files', function () {
-		assert.file(helperPath + "_grunt/styleguide.js");
+		assert.file(helperPath + "_grunt/postcssSeparator.js");
 	});
 
-	it('adds styleguide template', function () {
-		assert.file(helperPath + "templates/styleguide-template/index.html");
+	it('adds task to Gruntfile.js file', function () {
+		assert.fileContent("Gruntfile.js", /\'postcssSeparator\'/);
 	});
-
-	it('adds task to concurrent.js file', function () {
-		assert.fileContent(helperPath + "_grunt/concurrent.js", /styleguide/);
-	});
-
-	it('adds styleguide.md to sass folder', function () {
-		assert.file(srcPath + "scss/styleguide.md");
-	});
-
 });